requirements.txt
streamlit==1.31.1
crewai==1.9.3
python-dotenv==1.0.1
openpyxl==3.1.2
pdfplumber==0.11.0
litellm==1.43.12
together==1.2.3
pydantic==2.6.4
